Removed CopulaImplementation
