+++ lib/src/Uncertainty/Distribution/openturns/ClaytonCopula.hxx
@@ -83,15 +83,15 @@ public:
   Point computeCDFGradient(const Point & point) const override;
 
   /** Get the quantile of the distribution */
-  using DistributionImplementation::computeQuantile;
+  using ArchimedeanCopula::computeQuantile;
   Point computeQuantile(const Scalar prob, const Bool tail = false) const override;
 
   /** Compute the CDF of Xi | X1, ..., Xi-1. x = Xi, y = (X1,...,Xi-1) */
-  using CopulaImplementation::computeConditionalCDF;
+  using ArchimedeanCopula::computeConditionalCDF;
   Scalar computeConditionalCDF(const Scalar x, const Point & y) const override;
 
   /** Compute the quantile of Xi | X1, ..., Xi-1, i.e. x such that CDF(x|y) = q with x = Xi, y = (X1,...,Xi-1) */
-  using CopulaImplementation::computeConditionalQuantile;
+  using ArchimedeanCopula::computeConditionalQuantile;
   Scalar computeConditionalQuantile(const Scalar q, const Point & y) const override;
 
   /** Compute the archimedean generator of the archimedean copula, i.e.
@@ -135,9 +135,6 @@ protected:
 
 
 private:
-
-  /** Compute the covariance of the distribution */
-  void computeCovariance() const override;
 
   /** The parameter of the ClaytonCopula distribution */
   Scalar theta_;
