+++ lib/src/Uncertainty/Model/CMakeLists.txt
@@ -8,7 +8,6 @@ ot_add_source_file (DistributionImplementation.cxx)
 ot_add_source_file (ContinuousDistribution.cxx)
 ot_add_source_file (DiscreteDistribution.cxx)
 ot_add_source_file (EllipticalDistribution.cxx)
-ot_add_source_file (CopulaImplementation.cxx)
 ot_add_source_file (SklarCopula.cxx)
 ot_add_source_file (ArchimedeanCopula.cxx)
 ot_add_source_file (RandomVector.cxx)
@@ -56,7 +55,6 @@ ot_install_header_file (ThresholdEventImplementation.hxx)
 ot_install_header_file (ProcessEvent.hxx)
 ot_install_header_file (IntersectionEvent.hxx)
 ot_install_header_file (ArchimedeanCopula.hxx)
-ot_install_header_file (CopulaImplementation.hxx)
 ot_install_header_file (DistributionFactoryImplementation.hxx)
 ot_install_header_file (DiscreteDistribution.hxx)
 ot_install_header_file (ConstantRandomVector.hxx)
